Invert captcha dispatcher so routes keep type-hinting CaptchaInterface $captcha
Per PR #613 review: the original Captcha class was Google-specific and sat
behind the generic CaptchaInterface $captcha param, which made the whole
dispatch mechanism depend on renaming the param in every consumer.

Move the Google reCAPTCHA implementation into Recaptcha.php and reuse
Captcha.php as the dispatcher. Autowiring now resolves CaptchaInterface
$captcha to the dispatcher across every integration route and third-party
extension transparently — no consumer changes.

Settings follow the same inversion:
- SettingsCaptcha (new): provider selector + master toggle, claims the
  es_forms_settings_global_captcha filter slot.
- SettingsRecaptcha (was SettingsCaptcha): Google-only settings, now under
  the es_forms_settings_global_recaptcha filter.
- SettingsCaptchaProvider deleted; its responsibilities live on SettingsCaptcha.
