+++ src/Captcha/SettingsCaptcha.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Captcha Settings class - Google reCaptcha.
+ * Captcha provider settings wrapper — owns the single "Captcha" entry in Advanced.
  *
  * @package EightshiftForms\Captcha
  */
@@ -10,15 +10,15 @@ declare(strict_types=1);
 
 namespace EightshiftForms\Captcha;
 
-use EightshiftForms\Helpers\SettingsOutputHelpers;
-use EightshiftForms\Hooks\Variables;
 use EightshiftForms\Helpers\SettingsHelpers;
-use EightshiftForms\Labels\LabelsInterface;
+use EightshiftForms\Helpers\SettingsOutputHelpers;
 use EightshiftForms\Settings\SettingGlobalInterface;
 use EightshiftFormsVendor\EightshiftLibs\Services\ServiceInterface;
 
 /**
- * SettingsCaptcha class.
+ * Wraps the per-provider settings pages behind a single captcha menu entry
+ * with a provider selector. Routes the admin UI and runtime validity checks
+ * through whichever provider is currently active.
  */
 class SettingsCaptcha implements SettingGlobalInterface, ServiceInterface
 {
@@ -28,95 +28,28 @@ class SettingsCaptcha implements SettingGlobalInterface, ServiceInterface
 	public const FILTER_SETTINGS_GLOBAL_NAME = 'es_forms_settings_global_captcha';
 
 	/**
-	 * Filter settings global is Valid key.
+	 * Filter global settings is-valid key.
 	 */
 	public const FILTER_SETTINGS_GLOBAL_IS_VALID_NAME = 'es_forms_settings_global_is_valid_captcha';
 
 	/**
-	 * Settings key.
-	 */
-	public const SETTINGS_TYPE_KEY = 'captcha';
-
-	/**
-	 * Captcha Use key.
+	 * Master use key (shared across all providers).
 	 */
 	public const SETTINGS_CAPTCHA_USE_KEY = 'captcha-use';
 
 	/**
-	 * Google reCaptcha site key.
+	 * Provider option key.
 	 */
-	public const SETTINGS_CAPTCHA_SITE_KEY = 'captcha-site-key';
+	public const SETTINGS_CAPTCHA_PROVIDER_KEY = 'captcha-provider';
 
 	/**
-	 * Google reCaptcha secret key.
+	 * Provider identifiers.
 	 */
-	public const SETTINGS_CAPTCHA_SECRET_KEY = 'captcha-secret-key';
+	public const PROVIDER_GOOGLE = 'google';
+	public const PROVIDER_FRIENDLY = 'friendly';
 
 	/**
-	 * Google reCaptcha project_id key.
-	 */
-	public const SETTINGS_CAPTCHA_PROJECT_ID_KEY = 'captcha-project-id-key';
-	/**
-	 * Google reCaptcha api_key key.
-	 */
-	public const SETTINGS_CAPTCHA_API_KEY = 'captcha-api-key-key';
-
-	/**
-	 * Google reCaptcha score key.
-	 */
-	public const SETTINGS_CAPTCHA_SCORE_KEY = 'captcha-score';
-
-	/**
-	 * Google reCaptcha submit action key.
-	 */
-	public const SETTINGS_CAPTCHA_SUBMIT_ACTION_KEY = 'captcha-submit-action';
-	public const SETTINGS_CAPTCHA_SUBMIT_ACTION_DEFAULT_KEY = 'submit';
-
-	/**
-	 * Google reCaptcha score default key.
-	 */
-	public const SETTINGS_CAPTCHA_SCORE_DEFAULT_KEY = 0.5;
-
-	/**
-	 * Is enterprise version key.
-	 */
-	public const SETTINGS_CAPTCHA_ENTERPRISE_KEY = 'captcha-enterprise';
-
-	/**
-	 * Load on init key.
-	 */
-	public const SETTINGS_CAPTCHA_LOAD_ON_INIT_KEY = 'captcha-load-on-init';
-
-	/**
-	 * Google reCaptcha init action key.
-	 */
-	public const SETTINGS_CAPTCHA_INIT_ACTION_KEY = 'captcha-init-action';
-	public const SETTINGS_CAPTCHA_INIT_ACTION_DEFAULT_KEY = 'homepage';
-
-	/**
-	 * Hide badge key.
-	 */
-	public const SETTINGS_CAPTCHA_HIDE_BADGE_KEY = 'captcha-hide-badge';
-
-	/**
-	 * Instance variable for labels data.
-	 *
-	 * @var LabelsInterface
-	 */
-	protected $labels;
-
-	/**
-	 * Create a new instance.
-	 *
-	 * @param LabelsInterface $labels Inject documentsData which holds labels data.
-	 */
-	public function __construct(LabelsInterface $labels)
-	{
-		$this->labels = $labels;
-	}
-
-	/**
-	 * Register all the hooks
+	 * Register all the hooks.
 	 *
 	 * @return void
 	 */
@@ -127,42 +60,36 @@ class SettingsCaptcha implements SettingGlobalInterface, ServiceInterface
 	}
 
 	/**
-	 * Determine if settings global are valid.
+	 * Resolve the provider currently selected by the admin.
 	 *
-	 * @return boolean
+	 * Falls back to Google so existing installs (which only have `captcha-use`
+	 * set) keep working untouched after upgrade.
+	 *
+	 * @return string Provider identifier — either `google` or `friendly`.
 	 */
-	public function isSettingsGlobalValid(): bool
+	public static function getActiveProvider(): string
 	{
-		if (SettingsCaptchaProvider::getActiveProvider() !== SettingsCaptchaProvider::PROVIDER_GOOGLE) {
-			return false;
-		}
+		$value = SettingsHelpers::getOptionValue(self::SETTINGS_CAPTCHA_PROVIDER_KEY);
 
-		$isUsed = SettingsHelpers::isOptionCheckboxChecked(self::SETTINGS_CAPTCHA_USE_KEY, self::SETTINGS_CAPTCHA_USE_KEY);
-		$siteKey = (bool) SettingsHelpers::getOptionWithConstant(Variables::getGoogleReCaptchaSiteKey(), self::SETTINGS_CAPTCHA_SITE_KEY);
-		$secretKey = (bool) SettingsHelpers::getOptionWithConstant(Variables::getGoogleReCaptchaSecretKey(), self::SETTINGS_CAPTCHA_SECRET_KEY);
-		$apiKey = (bool) SettingsHelpers::getOptionWithConstant(Variables::getGoogleReCaptchaApiKey(), self::SETTINGS_CAPTCHA_API_KEY);
-		$projectIdKey = (bool) SettingsHelpers::getOptionWithConstant(Variables::getGoogleReCaptchaProjectIdKey(), self::SETTINGS_CAPTCHA_PROJECT_ID_KEY);
-
-		$isEnterprise = SettingsHelpers::isOptionCheckboxChecked(self::SETTINGS_CAPTCHA_ENTERPRISE_KEY, self::SETTINGS_CAPTCHA_ENTERPRISE_KEY);
-
-		if ($isEnterprise) {
-			if (!$isUsed || !$siteKey || !$apiKey || !$projectIdKey) {
-				return false;
-			}
-		} else {
-			if (!$isUsed || !$siteKey || !$secretKey) {
-				return false;
-			}
-		}
-
-		return true;
+		return $value === self::PROVIDER_FRIENDLY ? self::PROVIDER_FRIENDLY : self::PROVIDER_GOOGLE;
 	}
 
 	/**
-	 * Get global settings array for building settings page.
+	 * The merged page is valid whenever the active provider's own validity filter says so.
 	 *
-	 * Retained for BC in case the `FILTER_SETTINGS_GLOBAL_NAME` filter is still
-	 * consulted; the menu entry is now handled by `SettingsCaptchaProvider`.
+	 * @return bool
+	 */
+	public function isSettingsGlobalValid(): bool
+	{
+		$providerFilter = self::getActiveProvider() === self::PROVIDER_FRIENDLY
+			? SettingsFriendlyCaptcha::FILTER_SETTINGS_GLOBAL_IS_VALID_NAME
+			: SettingsRecaptcha::FILTER_SETTINGS_GLOBAL_IS_VALID_NAME;
+
+		return (bool) \apply_filters($providerFilter, false);
+	}
+
+	/**
+	 * Build the merged settings page.
 	 *
 	 * @return array<int, array<mixed>>
 	 */
@@ -172,206 +99,49 @@ class SettingsCaptcha implements SettingGlobalInterface, ServiceInterface
 			return SettingsOutputHelpers::getNoActiveFeature();
 		}
 
-		return [
-			SettingsOutputHelpers::getIntro(self::SETTINGS_TYPE_KEY),
+		$provider = self::getActiveProvider();
+
+		$output = [
+			SettingsOutputHelpers::getIntro('captcha'),
 			[
+				'component' => 'select',
+				'selectName' => SettingsHelpers::getOptionName(self::SETTINGS_CAPTCHA_PROVIDER_KEY),
+				'selectFieldLabel' => \__('Provider', 'eightshift-forms'),
+				// phpcs:ignore WordPress.WP.I18n.NoHtmlWrappedStrings
+				'selectFieldHelp' => \__('Pick which captcha service validates submissions. Switching the provider reloads the fields below.', 'eightshift-forms'),
+				'selectSingleSubmit' => true,
+				'selectContent' => [
+					[
+						'component' => 'select-option',
+						'selectOptionLabel' => \__('Google reCAPTCHA', 'eightshift-forms'),
+						'selectOptionValue' => self::PROVIDER_GOOGLE,
+						'selectOptionIsSelected' => $provider === self::PROVIDER_GOOGLE,
+					],
+					[
+						'component' => 'select-option',
+						'selectOptionLabel' => \__('Friendly Captcha', 'eightshift-forms'),
+						'selectOptionValue' => self::PROVIDER_FRIENDLY,
+						'selectOptionIsSelected' => $provider === self::PROVIDER_FRIENDLY,
+					],
+				],
+			],
+			[
+				'component' => 'divider',
+				'dividerExtraVSpacing' => true,
+			],
+		];
+
+		$providerTabs = $provider === self::PROVIDER_FRIENDLY
+			? SettingsFriendlyCaptcha::getProviderTabs()
+			: SettingsRecaptcha::getProviderTabs();
+
+		if ($providerTabs) {
+			$output[] = [
 				'component' => 'tabs',
-				'tabsContent' => self::getProviderTabs(),
-			],
-		];
-	}
+				'tabsContent' => $providerTabs,
+			];
+		}
 
-	/**
-	 * Tab definitions for the Google reCAPTCHA provider, composed into the
-	 * merged captcha page by `SettingsCaptchaProvider`.
-	 *
-	 * @return array<int, array<string, mixed>>
-	 */
-	public static function getProviderTabs(): array
-	{
-		$isEnterprise = SettingsHelpers::isOptionCheckboxChecked(self::SETTINGS_CAPTCHA_ENTERPRISE_KEY, self::SETTINGS_CAPTCHA_ENTERPRISE_KEY);
-		$isInit = SettingsHelpers::isOptionCheckboxChecked(self::SETTINGS_CAPTCHA_LOAD_ON_INIT_KEY, self::SETTINGS_CAPTCHA_LOAD_ON_INIT_KEY);
-
-		return [
-			[
-				'component' => 'tab',
-				'tabLabel' => \__('General', 'eightshift-forms'),
-				'tabContent' => [
-					[
-						'component' => 'intro',
-						// phpcs:ignore WordPress.WP.I18n.NoHtmlWrappedStrings
-						'introSubtitle' => \__('Protect your website from spam and abuse using Google\'s reCAPTCHA.<br />A captcha is a simple task that is easy for humans to do, but difficult for bots.', 'eightshift-forms'),
-					],
-					[
-						'component' => 'checkboxes',
-						'checkboxesFieldHideLabel' => true,
-						'checkboxesName' => SettingsHelpers::getOptionName(self::SETTINGS_CAPTCHA_ENTERPRISE_KEY),
-						'checkboxesContent' => [
-							[
-								'component' => 'checkbox',
-								'checkboxLabel' => \__('Use reCAPTCHA Enterprise', 'eightshift-forms'),
-								'checkboxIsChecked' => $isEnterprise,
-								'checkboxValue' => self::SETTINGS_CAPTCHA_ENTERPRISE_KEY,
-								'checkboxSingleSubmit' => true,
-								'checkboxAsToggle' => true,
-							],
-						],
-					],
-					[
-						'component' => 'divider',
-						'dividerExtraVSpacing' => true,
-					],
-					SettingsOutputHelpers::getPasswordFieldWithGlobalVariable(
-						Variables::getGoogleReCaptchaSiteKey(),
-						self::SETTINGS_CAPTCHA_SITE_KEY,
-						'ES_GOOGLE_RECAPTCHA_SITE_KEY',
-						\__('Site key', 'eightshift-forms'),
-					),
-
-					...(!$isEnterprise ? [
-						SettingsOutputHelpers::getPasswordFieldWithGlobalVariable(
-							Variables::getGoogleReCaptchaSecretKey(),
-							self::SETTINGS_CAPTCHA_SECRET_KEY,
-							'ES_GOOGLE_RECAPTCHA_SECRET_KEY',
-							\__('Secret key', 'eightshift-forms'),
-						),
-					] : [
-						SettingsOutputHelpers::getInputFieldWithGlobalVariable(
-							Variables::getGoogleReCaptchaProjectIdKey(),
-							self::SETTINGS_CAPTCHA_PROJECT_ID_KEY,
-							'ES_GOOGLE_RECAPTCHA_PROJECT_ID_KEY',
-							\__('Project ID', 'eightshift-forms'),
-						),
-						SettingsOutputHelpers::getPasswordFieldWithGlobalVariable(
-							Variables::getGoogleReCaptchaApiKey(),
-							self::SETTINGS_CAPTCHA_API_KEY,
-							'ES_GOOGLE_RECAPTCHA_API_KEY',
-							\__('API key', 'eightshift-forms'),
-						),
-					]),
-				],
-			],
-			[
-				'component' => 'tab',
-				'tabLabel' => \__('Advanced', 'eightshift-forms'),
-				'tabContent' => [
-					[
-						'component' => 'checkboxes',
-						'checkboxesFieldHideLabel' => true,
-						'checkboxesName' => SettingsHelpers::getOptionName(self::SETTINGS_CAPTCHA_HIDE_BADGE_KEY),
-						'checkboxesContent' => [
-							[
-								'component' => 'checkbox',
-								'checkboxLabel' => \__('Hide badge', 'eightshift-forms'),
-								'checkboxIsChecked' => SettingsHelpers::isOptionCheckboxChecked(self::SETTINGS_CAPTCHA_HIDE_BADGE_KEY, self::SETTINGS_CAPTCHA_HIDE_BADGE_KEY),
-								'checkboxValue' => self::SETTINGS_CAPTCHA_HIDE_BADGE_KEY,
-								'checkboxSingleSubmit' => true,
-								'checkboxAsToggle' => true,
-								'checkboxHelp' => \__('Not recommended, as it is against Google\'s terms of use.', 'eightshift-forms'),
-							],
-						],
-					],
-					[
-						'component' => 'divider',
-						'dividerExtraVSpacing' => true,
-					],
-					[
-						'component' => 'input',
-						'inputName' => SettingsHelpers::getOptionName(self::SETTINGS_CAPTCHA_SCORE_KEY),
-						'inputFieldLabel' => \__('"Spam unlikely" threshold', 'eightshift-forms'),
-						'inputFieldHelp' => \__('The level above which a submission is <strong>not</strong> considered spam. Should be between 0.1 and 1.0.<br />In most cases, a user will receive as core between 0.8 and 0.9.', 'eightshift-forms'),
-						'inputType' => 'number',
-						'inputValue' => SettingsHelpers::getOptionValue(self::SETTINGS_CAPTCHA_SCORE_KEY),
-						'inputMin' => 0.1,
-						'inputMax' => 1,
-						'inputStep' => 0.1,
-						'inputIsNumber' => true,
-						'inputPlaceholder' => self::SETTINGS_CAPTCHA_SCORE_DEFAULT_KEY,
-					],
-					[
-						'component' => 'divider',
-						'dividerExtraVSpacing' => true,
-					],
-					[
-						'component' => 'input',
-						'inputName' => SettingsHelpers::getOptionName(self::SETTINGS_CAPTCHA_SUBMIT_ACTION_KEY),
-						'inputFieldLabel' => \__('"On submit" action name', 'eightshift-forms'),
-						'inputFieldHelp' => \__('Name of the action sent to reCAPTCHA on form submission.', 'eightshift-forms'),
-						'inputType' => 'text',
-						'inputValue' => SettingsHelpers::getOptionValue(self::SETTINGS_CAPTCHA_SUBMIT_ACTION_KEY),
-						'inputPlaceholder' => self::SETTINGS_CAPTCHA_SUBMIT_ACTION_DEFAULT_KEY,
-					],
-					[
-						'component' => 'divider',
-						'dividerExtraVSpacing' => true,
-					],
-					[
-						'component' => 'checkboxes',
-						'checkboxesFieldHideLabel' => true,
-						'checkboxesName' => SettingsHelpers::getOptionName(self::SETTINGS_CAPTCHA_LOAD_ON_INIT_KEY),
-						'checkboxesContent' => [
-							[
-								'component' => 'checkbox',
-								'checkboxLabel' => \__('Load Captcha on website load', 'eightshift-forms'),
-								'checkboxIsChecked' => SettingsHelpers::isOptionCheckboxChecked(self::SETTINGS_CAPTCHA_LOAD_ON_INIT_KEY, self::SETTINGS_CAPTCHA_LOAD_ON_INIT_KEY),
-								'checkboxValue' => self::SETTINGS_CAPTCHA_LOAD_ON_INIT_KEY,
-								'checkboxHelp' => \__('By default, Captcha is only loaded on pages that contain forms. However, with this option, you can load Captcha on every page.', 'eightshift-forms'),
-								'checkboxSingleSubmit' => true,
-								'checkboxAsToggle' => true,
-								'checkboxAsToggleSize' => 'medium',
-							],
-						],
-					],
-					$isInit ? [
-						'component' => 'input',
-						'inputName' => SettingsHelpers::getOptionName(self::SETTINGS_CAPTCHA_INIT_ACTION_KEY),
-						'inputFieldLabel' => \__('Action name', 'eightshift-forms'),
-						'inputFieldHelp' => \__('Name of the action sent to reCAPTCHA when Captcha is loaded on every page.', 'eightshift-forms'),
-						'inputType' => 'text',
-						'inputValue' => SettingsHelpers::getOptionValue(self::SETTINGS_CAPTCHA_INIT_ACTION_KEY),
-						'inputPlaceholder' => self::SETTINGS_CAPTCHA_INIT_ACTION_DEFAULT_KEY,
-					] : [],
-				],
-			],
-			[
-				'component' => 'tab',
-				'tabLabel' => \__('Help', 'eightshift-forms'),
-				'tabContent' => [
-					[
-						'component' => 'steps',
-						'stepsTitle' => \__('How to get the Free reCAPTCHA API key?', 'eightshift-forms'),
-						'stepsContent' => [
-							// translators: %s will be replaced with the external link.
-							\sprintf(\__('Visit this <a href="%s" target="_blank" rel="noopener noreferrer">link</a>.', 'eightshift-forms'), 'https://www.google.com/recaptcha/admin/create'),
-							\__('Configure all the options. Make sure to select <strong>reCaptcha version 3</strong>!', 'eightshift-forms'),
-							\__('Copy the API key into the field under the API tab or use the global constant.', 'eightshift-forms'),
-						],
-					],
-					[
-						'component' => 'divider',
-						'dividerExtraVSpacing' => true,
-					],
-					[
-						'component' => 'steps',
-						'stepsTitle' => \__('How to get the Enterprise reCAPTCHA API key?', 'eightshift-forms'),
-						'stepsContent' => [
-							// translators: %s will be replaced with the external link.
-							\sprintf(\__('Visit <a href="%s" target="_blank" rel="noopener noreferrer">Google Cloud Console</a>.', 'eightshift-forms'), 'https://console.cloud.google.com/'),
-							\__('Create a new project and set that project as <strong>Project ID</strong>.', 'eightshift-forms'),
-							\__('Search and go to <strong>reCAPTCHA</strong> product.', 'eightshift-forms'),
-							\__('You will probably need to set billing service for this product.', 'eightshift-forms'),
-							\__('Create a new key and set that key as <strong>Site key</strong>.', 'eightshift-forms'),
-							// translators: %s will be replaced with the website domain.
-							\sprintf(\__('Limit the key to your website domain. Domain: <strong>%s</strong> (exact, no trailing slash and protocol).', 'eightshift-forms'), \preg_replace("(^https?://)", "", \site_url())),
-							\__('Search and go to <strong>API & Services</strong> product.', 'eightshift-forms'),
-							\__('Go to <strong>Credentials</strong> section and create a new API key.', 'eightshift-forms'),
-							// translators: %s will be replaced with the website domain.
-							\sprintf(\__('Create a new key for <strong>Website</strong>, add restrictions to your website domain <strong>%s</strong> (exact, no trailing slash, with protocol) and set API restrictions to <strong>reCAPTCHA Enterprise</strong>.', 'eightshift-forms'), \esc_url(\site_url())),
-							\__('Set that key as <strong>API key</strong>.', 'eightshift-forms'),
-						],
-					],
-				],
-			],
-		];
+		return $output;
 	}
 }
