+++ src/Rest/Routes/Integrations/ActiveCampaign/FormSubmitActiveCampaignRoute.php
@@ -50,7 +50,7 @@ class FormSubmitActiveCampaignRoute extends AbstractIntegrationFormSubmit
 	 * @param SecurityInterface $security Inject security methods.
 	 * @param ValidatorInterface $validator Inject validator methods.
 	 * @param LabelsInterface $labels Inject labels methods.
-	 * @param CaptchaInterface $captchaDispatcher Inject captcha methods.
+	 * @param CaptchaInterface $captcha Inject captcha methods.
 	 * @param MailerInterface $mailer Inject mailerInterface methods.
 	 * @param EnrichmentInterface $enrichment Inject enrichment methods.
 	 * @param ActiveCampaignClientInterface $activeCampaignClient Inject ActiveCampaign methods.
@@ -59,7 +59,7 @@ class FormSubmitActiveCampaignRoute extends AbstractIntegrationFormSubmit
 		SecurityInterface $security,
 		ValidatorInterface $validator,
 		LabelsInterface $labels,
-		CaptchaInterface $captchaDispatcher,
+		CaptchaInterface $captcha,
 		MailerInterface $mailer,
 		EnrichmentInterface $enrichment,
 		ActiveCampaignClientInterface $activeCampaignClient
@@ -67,7 +67,7 @@ class FormSubmitActiveCampaignRoute extends AbstractIntegrationFormSubmit
 		$this->security = $security;
 		$this->validator = $validator;
 		$this->labels = $labels;
-		$this->captcha = $captchaDispatcher;
+		$this->captcha = $captcha;
 		$this->mailer = $mailer;
 		$this->enrichment = $enrichment;
 		$this->activeCampaignClient = $activeCampaignClient;
