+++ src/Rest/Routes/Integrations/Mailerlite/FormSubmitMailerliteRoute.php
@@ -49,7 +49,7 @@ class FormSubmitMailerliteRoute extends AbstractIntegrationFormSubmit
 	 * @param SecurityInterface $security Inject security methods.
 	 * @param ValidatorInterface $validator Inject validator methods.
 	 * @param LabelsInterface $labels Inject labels methods.
-	 * @param CaptchaInterface $captchaDispatcher Inject captcha methods.
+	 * @param CaptchaInterface $captcha Inject captcha methods.
 	 * @param MailerInterface $mailer Inject mailerInterface methods.
 	 * @param EnrichmentInterface $enrichment Inject enrichment methods.
 	 * @param ClientInterface $mailerliteClient Inject mailerliteClient methods.
@@ -58,7 +58,7 @@ class FormSubmitMailerliteRoute extends AbstractIntegrationFormSubmit
 		SecurityInterface $security,
 		ValidatorInterface $validator,
 		LabelsInterface $labels,
-		CaptchaInterface $captchaDispatcher,
+		CaptchaInterface $captcha,
 		MailerInterface $mailer,
 		EnrichmentInterface $enrichment,
 		ClientInterface $mailerliteClient
@@ -66,7 +66,7 @@ class FormSubmitMailerliteRoute extends AbstractIntegrationFormSubmit
 		$this->security = $security;
 		$this->validator = $validator;
 		$this->labels = $labels;
-		$this->captcha = $captchaDispatcher;
+		$this->captcha = $captcha;
 		$this->mailer = $mailer;
 		$this->enrichment = $enrichment;
 		$this->mailerliteClient = $mailerliteClient;
